requirements.txt
absl-py==0.3.0
APScheduler==3.5.1
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.1.5
attrs==18.1.0
Automat==0.7.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
baidu-aip==2.2.5.2
beautifulsoup4==4.6.3
bleach==1.5.0
boto3==1.5.20
botocore==1.8.50
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==5.1
cloudpickle==0.5.2
colorama==0.3.9
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
cymem==1.31.2
#cytoolz==0.9.0.1
DBUtils==1.3
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
duckling==1.8.0
enum34==1.1.6
fakeredis==0.10.3
fbmessenger==5.2.0
Flask==1.0.2
Flask-Cors==3.0.4
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gast==0.2.0
gevent==1.2.2
graphviz==0.8.3
greenlet==0.4.14
grpcio==1.14.0
h5py==2.7.1
html5lib==0.9999999
humanfriendly==4.16.1
hyperlink==17.3.1
idna==2.6
imgkit==1.0.1
incremental==17.5.0
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
itchat==1.2.32
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.2
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
#Keras==2.1.6
#Keras-Applications==1.0.4
#Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
klein==17.10.0
logger==1.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mattermostwrapper==2.1
#mitie==0.5.0
mock==2.0.0
monotonic==1.5
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.1
msgpack-python==0.5.4
murmurhash==0.28.0
neo4j==1.7.0b1
neo4j-driver==1.6.1
neobolt==1.7.0b1
neotime==1.0.0
networkx==2.1
numpy==1.14.4
packaging==17.1
pandas==0.23.4
pathlib==1.0.1
pathlib2==2.3.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
#pkg-resources==0.0.0pkg-resources==0.0.0
plac==0.9.6
pluggy==0.7.1
preshed==1.0.1
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
py==1.5.4
#py2neo==4.1.0
pycparser==2.18
pydub==0.22.1
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pykwalify==1.6.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
pystack-debugger==0.8.0
pytest==3.7.2
python-crfsuite==0.9.6
python-dateutil==2.7.3
python-telegram-bot==10.1.0
pytz==2018.5
PyYAML==3.12
pyzmq==17.1.2
qqbot==2.3.10
#rasa-core==0.10.3
#rasa-nlu==0.13.0
redis==2.10.6
regex==2017.4.5
requests==2.18.4
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.37
s3transfer==0.1.13
scandir==1.9.0
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
slackclient==1.2.1
#spacy==2.0.11
subprocess32==3.5.2
tabulate==0.8.2
#tensorboard==1.8.0
#tensorflow==1.8.0
termcolor==1.1.0
#thinc==6.10.3
toolz==0.9.0
tornado==5.1
tqdm==4.19.5
traitlets==4.3.2
twilio==6.14.0
Twisted==18.7.0
typing==3.6.2
tzlocal==1.5.1
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
wrapt==1.10.11
wxpy==0.3.9.8
zope.interface==4.5.0
MySQL-python==1.2.5
DBUtils==1.3
baidu-aip==2.2.5.2